replace filters with tabs on bridges risk view
